refactor(tab): extract tab-panels component to a single folder
