+++ src/components/TabPanels/tab-panels.scss
@@ -1,0 +1,7 @@
+.sb-tab-panel {
+  display: block;
+
+  &[aria-hidden="true"] {
+    display: none;
+  }
+}
