+++ src/components/Tabs/components/index.js
@@ -1,2 +1,1 @@
-export * from './panel'
 export * from './tab'
